+++ services/article/api/readmodel/src/main/java/nettee/blolet/article/readmodel/ArticleQueryModels.java
@@ -1,4 +1,4 @@
-package nettee.article.readmodel;
+package nettee.blolet.article.readmodel;
 
 import lombok.Builder;
 import nettee.blolet.article.domain.sub.ArticleStatus;
