+++ services/article/application/src/main/java/nettee/article/port/ArticleQueryRepositoryPort.java
@@ -1,7 +1,7 @@
 package nettee.article.port;
 
-import nettee.article.readmodel.ArticleQueryModels.ArticleDetail;
-import nettee.article.readmodel.ArticleQueryModels.ArticleSummary;
+import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleDetail;
+import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleSummary;
 import org.springframework.data.domain.Slice;
 
 import java.time.Instant;
