+++ services/article/driven/rdb/src/main/java/nettee/article/driven/rdb/persistence/ArticleQueryAdapter.java
@@ -3,8 +3,8 @@ package nettee.article.driven.rdb.persistence;
 import lombok.RequiredArgsConstructor;
 import nettee.article.driven.rdb.persistence.mapper.ArticleEntityMapper;
 import nettee.article.port.ArticleQueryRepositoryPort;
-import nettee.article.readmodel.ArticleQueryModels.ArticleDetail;
-import nettee.article.readmodel.ArticleQueryModels.ArticleSummary;
+import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleDetail;
+import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleSummary;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
