+++ services/article/driven/rdb/src/main/java/nettee/article/driven/rdb/persistence/mapper/ArticleEntityMapper.java
@@ -4,8 +4,8 @@ import nettee.blolet.article.domain.Article;
 import nettee.article.driven.rdb.entity.ArticleEntity;
 import nettee.article.driven.rdb.projection.ArticleQueryProjections.ArticleDetailProjection;
 import nettee.article.driven.rdb.projection.ArticleQueryProjections.ArticleSummaryProjection;
-import nettee.article.readmodel.ArticleQueryModels.ArticleDetail;
-import nettee.article.readmodel.ArticleQueryModels.ArticleSummary;
+import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleDetail;
+import nettee.blolet.article.readmodel.ArticleQueryModels.ArticleSummary;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
